style : #78 상단 네비게이션 수정
