+++ Spoony-iOS/Spoony-iOS/Source/Feature/Search/Models/SearchResult.swift
