+++ Spoony-iOS/Spoony-iOS/Source/Feature/Search/SearchState.swift
